change construction of sulfur/ui/model#object to on-demand
The automatic construction of #object after each model change (also
caused by changes published by an attribute value) happens to introduce
a race condition when used in asynchronous code, because the publisher
works asynchronously.
The simplest solution is to change the object construction into an
on-demand approach, which should have no impact on performance.